log.hpp and log.cpp created
