+++ include/http.hpp
@@ -7,7 +7,3 @@ void SetupHttpHandlers();
 
 extern WebServer configServer;
 extern IotWebConf iotWebConf;
-
-#if defined(EBUS_INTERNAL)
-void addLog(String entry);
-#endif
