+++ include/log.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <Arduino.h>
+
+void addLog(String entry);
+String getLog();
